autoload functions from ~/.zsh/functions
It is a fairly standard practice to autoload functions in ZSH. When an
autoloaded function is executed for the first time, ZSH searches for a
file with the same name in the fpath environment variable and executes
the content of that file as the method body.

With this commit, any function defined in ~/.zsh/functions is
automatically exported to the environment as an autoloaded function. A
step towards modular ZSH configuration!